fix(observe): PE review quick wins — hints in summaries, pure hint fn, timeline extraction
1. Add diagnostic hints to summary mode responses for network_bodies and
   websocket_events (P1) — previously the summary early-return bypassed
   hint injection when results were empty.

2. Decouple networkBodiesEmptyHint from Deps (P2) — now accepts
   waterfallCount as int, making the function pure and trivially testable.

3. Add unit tests for empty_hints.go (P2) — direct function-level tests
   in the observe package without the full test harness.

4. Extract timeline code from analysis.go into timeline.go (P1) —
   brings analysis.go from 793 to 602 LOC, well under the 800 limit.

5. Fix URL filter hint wording (P2) — changed "exist unfiltered" to
   "exist in the buffer" for accuracy.
